requirements.txt


# modelChack
# conda create -n modelChack python=3.10 -y

Flask==2.3.3
flask-cors==5.0.1
Flask-JWT-Extended==4.6.0
Flask-PyMongo==2.3.0
Werkzeug==2.3.7
itsdangerous==2.2.0
Jinja2==3.1.6
python-dotenv==1.0.1
PyJWT==2.10.1
pymongo==4.12.1
qrcode

firebase-admin==6.4.0
google-api-core==2.25.0rc1
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.2
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
httplib2==0.22.0
proto-plus==1.26.1
protobuf==5.29.4
uritemplate==4.1.1
rsa==4.9.1



# AI

dlib==19.24.8
face-recognition==1.3.0
face_recognition_models==0.3.0
opencv-python==4.11.0.86
pillow==11.2.1
numpy==2.2.5

# ai support

requests==2.32.3
urllib3==2.4.0
certifi==2025.4.26
charset-normalizer==3.4.2
idna==3.10


# bullid tool

pip==25.1
setuptools==78.1.1
wheel==0.45.1

gunicorn

# # faceenv balace ai same  env = modelChack
# Flask
# Flask-PyMongo
# Flask-JWT-Extended
# python-dotenv
# firebase-admin

# qrcode

# bcrypt
# Werkzeug




# # photoharper project 

# # modelChack run env 